requirements.txt
anthropic[bedrock]>=0.40.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
boto3>=1.34.0
